Atualização Monstro

+++ MvcCalendario/src/MvcCalendario.Business/Services/ClienteService.cs
@@ -11,20 +11,30 @@ namespace MvcCalendario.Business.Services
     {
         private readonly IClienteRepository _clienteRepository;
         private readonly IEnderecoRepository _enderecoRepository;
+        private readonly IContatoRepository _contatoRepository;
+
 
         public ClienteService(IClienteRepository ClienteRepository,
                                  IEnderecoRepository enderecoRepository,
+                                 IContatoRepository contatoRepository,
                                  INotificador notificador) : base(notificador)
         {
             _clienteRepository = ClienteRepository;
             _enderecoRepository = enderecoRepository;
+            _contatoRepository = contatoRepository;
         }
 
-        public async Task Adicionar(Cliente Cliente)
+        public async Task Adicionar(Cliente cliente)
         {
-            if (!ExecutarValidacao(new ClienteValidation(), Cliente)) return;
+            if (!ExecutarValidacao(new ClienteValidation(), cliente)) return;
 
-            await _clienteRepository.Adicionar(Cliente);
+            if (_clienteRepository.Buscar(f => f.CPF == cliente.CPF).Result.Any())
+            {
+                Notificar("Já existe um cliente com este documento infomado.");
+                return;
+            }
+
+            await _clienteRepository.Adicionar(cliente);
         }
 
         public async Task Atualizar(Cliente Cliente)
@@ -36,20 +46,38 @@ namespace MvcCalendario.Business.Services
 
         public async Task Remover(Guid id)
         {
-            if (_clienteRepository.ObterPorId(id).Result.Enderecos.Any())
+            try
             {
-                Notificar("O Cliente possui endereços cadastrados!");
-                return;
+                var result = _clienteRepository.ObterClienteCompleto(id).Result;
+
+                if (result.Enderecos != null && result.Enderecos.Any())
+                {
+                    Notificar("O Cliente possui endereços cadastrados!");
+                    return;
+                }
+
+                if (result.Contatos != null && result.Contatos.Any())
+                {
+                    Notificar("O Cliente possui contatos cadastrados!");
+                    return;
+                }
+
+                var enderecos = _enderecoRepository.ObterEnderecosPorCliente(id);
+                var contatos = _contatoRepository.ObterContatosPorCliente(id);
+
+                if (enderecos != null) await _clienteRepository.RemoverEnderecosPorCliente(id);
+                if (contatos != null) await _clienteRepository.RemoverContatosPorCliente(id);
+
+                await _clienteRepository.Remover(id);
+
+            }
+            catch (Exception ex)
+            {
+
+                throw ex;
             }
 
-            var cliente = await _clienteRepository.ObterPorId(id);
 
-            if (cliente.Enderecos.Any())
-            {
-                await _clienteRepository.RemoverEnderecosPorCliente(id);
-            }
-
-            await _clienteRepository.Remover(id);
         }
